+++ Graphics/src/acceleration/Cell.java
@@ -2,15 +2,15 @@ package acceleration;
 
 public class Cell {
 
-	private double minX;
-	private double maxX;
-	private double minY;
-	private double maxY;
-	private double minZ;
-	private double maxZ;
+	private float minX;
+	private float maxX;
+	private float minY;
+	private float maxY;
+	private float minZ;
+	private float maxZ;
 	private int cellNumber;
 	
-	public Cell(double x, double d, double y, double e, double z, double f, int cellNumber) {
+	public Cell(float x, float d, float y, float e, float z, float f, int cellNumber) {
 		this.minX = x;
 		this.maxX = d;
 		this.minY = y;
@@ -20,51 +20,51 @@ public class Cell {
 		this.cellNumber = cellNumber;
 	}
 
-	public double getMinX() {
+	public float getMinX() {
 		return minX;
 	}
 
-	public void setMinX(double minX) {
+	public void setMinX(float minX) {
 		this.minX = minX;
 	}
 
-	public double getMaxX() {
+	public float getMaxX() {
 		return maxX;
 	}
 
-	public void setMaxX(double maxX) {
+	public void setMaxX(float maxX) {
 		this.maxX = maxX;
 	}
 
-	public double getMinY() {
+	public float getMinY() {
 		return minY;
 	}
 
-	public void setMinY(double minY) {
+	public void setMinY(float minY) {
 		this.minY = minY;
 	}
 
-	public double getMaxY() {
+	public float getMaxY() {
 		return maxY;
 	}
 
-	public void setMaxY(double maxY) {
+	public void setMaxY(float maxY) {
 		this.maxY = maxY;
 	}
 
-	public double getMinZ() {
+	public float getMinZ() {
 		return minZ;
 	}
 
-	public void setMinZ(double minZ) {
+	public void setMinZ(float minZ) {
 		this.minZ = minZ;
 	}
 
-	public double getMaxZ() {
+	public float getMaxZ() {
 		return maxZ;
 	}
 
-	public void setMaxZ(double maxZ) {
+	public void setMaxZ(float maxZ) {
 		this.maxZ = maxZ;
 	}
 }
